% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITDRMS.clean.R
\name{ITDRMS.clean}
\alias{ITDRMS.clean}
\title{ITDRMS.clean}
\usage{
<<<<<<< HEAD
ITDRMS.clean(data = NULL, keratins = TRUE, serum = TRUE, trypsin = TRUE)
=======
ITDRMS.clean(
  data = NULL,
  keratins = TRUE,
  serum = TRUE,
  trypsin = TRUE,
  inf.data = TRUE
)
>>>>>>> 2448b81 (Update)
}
\arguments{
\item{data}{Data frame with the loaded mass spec data}

\item{serum}{Logical: Should the serum proteins be removed?}

\item{trypsin}{Logical: Should trypsin be removed?}

\item{keratin}{Logical: Should the keratins be removed?}
}
\value{
List of two elements. $data is the data frame with contaminants removed and $removed is the data frame listing the removed contaminants.
}
\description{
Cleans the loaded mass spec data
}
\examples{
data_cleaned <- ITDRMS.clean(data_raw)
}
