% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITDRMS.fit.R
\name{ITDRMS.fit}
\alias{ITDRMS.fit}
\title{ITDRMS.fit}
\usage{
ITDRMS.fit(
  data = NULL,
  ncores = 1,
  ram = 8,
  fit.length = 100,
  control.slope = c("horizontal", "linear", "sigmoid"),
  outlier.removal = TRUE
)
}
\arguments{
\item{data}{Data frame scaled mass spec data}

\item{ncores}{Integer: How many cores to use for fitting. Default is 1.}

\item{ram}{Integer: Size allowed for parallel computing in GB.}

\item{fit.length}{Integer: How many points should be used for fitting curves. Default is 100 which is sufficient for plotting.}

\item{control.slope}{Character: "horizontal" for linear fit with no slope, "linear" for linear with with slope, "sigmoid" for LL4 fit.}

\item{outlier.removal}{Logical: If TRUE (default is TRUE), outliers will be identified as points that significantly worsen the fit and removed.}
}
\value{
A list with three elements. $data is a data frame with the fitted data, $fits is a list with fit objects and $curvy_controls is a data frame with those proteins, that show dose-response behaviour at the lowest temperature (baseline has a sigmoid shape).
}
\description{
Fits the scaled data.
}
\examples{
data_fitted <- ITDRMS.fit(data_scaled)
}
