% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITDRMS.load.R
\name{ITDRMS.load}
\alias{ITDRMS.load}
\title{ITDRMS.load}
\usage{
ITDRMS.load(
  datafolder = "datafiles",
  pattern = ".*",
  orgs = c("Hs", "Pf"),
  temperatures = c(37, 53, 59, 65),
  top.conc = 10,
  dilution = 4,
  direction = "decreasing",
  control.tube = "last",
  tubes = 10,
  concentrations = NA
)
}
\arguments{
\item{datafolder}{Directory with datafiles. The name of the files needs to the pattern (see below), temperature (e.g. _37_) and organism abbreviation (e.g. _Hs). Default is "datafiles"}

\item{pattern}{Character string that is contained in all filenames of files that are to be loaded. Default is ".*".}

\item{orgs}{Character vector listing files with data of what organisms there are in the datafolder. Default is c("Pf","Hs").}

\item{temperatures}{Integer vector with temperatures at which the assay was conducted. Default is c(37,53,59,65).}

\item{top.conc}{Integer. Top concentration of the drug dilution series.}

\item{direction}{Character string. "decreasing" when the loaded data contains highest concentration first, "increasing" otherwise.}

\item{control.tube}{Character string. "last" if the vehicle control is last in the loaded data, otherwise "first".}

\item{tubes}{Integer. How many tubes (including vehicle control)? Default is 10 assuming TMT10 labeling.}

\item{concentrations}{Integer vector with the concentrations of drug used in the assay. Default is NA. If specified, it overwrites
dilution, tp.conc, direction, control.tube and tubes.}

\item{dilutions}{Integer stating the dilution factor of the drug dilution series.}
}
\value{
List with two data.frames: $curves can be used in the next step of ITDRMS pipeline, $data can be saved as a csv.
}
\description{
Loads ITDR-CETSA datafiles from Proteome Discover (Versions 2.0 and above) to R. Protein abundances should be exported as .txt files and their filenames must
indicate organism and temperature, for example 'CETSAITDR_cipargamin_37_Pf.txt'.
}
\examples{
ITDRMS.load(temperatures=c(37,53,59,65))
}
