% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITDRMS.plot.R
\name{ITDRMS.plot}
\alias{ITDRMS.plot}
\title{ITDRMS.plot}
\usage{
ITDRMS.plot(
  data = NULL,
  fits = NULL,
  hits = NULL,
  calc.POS = NULL,
  POS.source = Pf_meltcurve_lysate,
  scale = FALSE,
  print.stats = FALSE,
  color.scheme = "rainbow",
  fit.length = 100,
  pdf.export = FALSE,
  pdf.folder = ".",
  pdf.name = "ITDR_curves"
)
}
\arguments{
\item{data}{Data frame: Scaled data with removed outliers and fitting statistics, ideally $data element from ITDRMS.fit output.}

\item{fits}{List containing all fit objects, ideally $fits element from ITDRMS.fit output.}

\item{hits}{Data frame with hits coming from ITDRMS.hit. Only needed when print.stats=TRUE}

\item{calc.POS}{vector: for which temperatures should POS be calculated? Default is NULL, in which case POS will not be calculated. Use with care, ideally only with a small subselection of proteins, such as identified hits.}

\item{POS.source}{Data frame with melting curve data to calculate POS from.}

\item{scale}{Logical: Should the curves for each protein be scaled from 0 to 1? Default is FALSE. If TRUE, the amplitude of response will not
be comparable across proteins!}

\item{print.stats}{Logical: Should the R2 and p-value be printed in the plots? Default is FALSE.}

\item{color.scheme}{Character string: "rainbow" will plot curves from red to violet from lowest to highest temperature, "distinct" will use distinct colors}

\item{fit.length}{Integer: How many points should be used for fitting curves. Default is 100 which is sufficient for plotting.}

\item{pdf.export}{Logical: If TRUE (default), a pdf with all plots will be exported.}

\item{pdf.folder}{Character string: Name of the directory for pdf export. Default is the working directory.}

\item{pdf.name}{Character string: Name of tthe exported pdf file. Default is 'ITDR_curves'}
}
\value{
A list of all plots.
}
\description{
Identifies hits from fitted mass spec data.
}
\examples{
data_plotted <- ITDRMS.plot(data_fitted$data, data_fitted$fits)
}
