% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITDRMS.POS.R
\name{ITDRMS.POS}
\alias{ITDRMS.POS}
\title{ITDRMS.plot}
\usage{
ITDRMS.POS(
  protein = NA,
  source.data = Pf_meltcurve_lysate,
  temperatures = c(50, 55, 60)
)
}
\arguments{
\item{protein}{Character string: Identifier of a protein that can be found in 'id' column of source.data.}

\item{source.data}{Data frame with measured melting curve, with one column for protein 'id' and other columns for scaled(zero to one) values}

\item{temperatures}{vector: for which temperatures should POS be calculated? Default is c(50,55,60)}
}
\value{
Data frame with calculated plausibility values.
}
\description{
Identifies hits from fitted mass spec data.
}
\examples{
plausibilities <- ITDRMS.plot("PF3D7_1017500", ITDRMS::source.data.lysate)
}
