% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITDRMS.scale.R
\name{ITDRMS.scale}
\alias{ITDRMS.scale}
\title{ITDRMS.scale}
\usage{
ITDRMS.scale(
  data = NULL,
  remove.control.outliers = TRUE,
  remove.hightemp.outliers = FALSE,
  normalization.points = 1,
  normalize.baseline = TRUE,
  normalize.selection = NULL
)
}
\arguments{
\item{data}{Data frame with the loaded and ideally cleaned up mass spec data}

\item{remove.control.outliers}{Logical. If TRUE (default), the program will remove baseline points that significantly deviate from a straight line fit.}

\item{remove.hightemp.outliers}{Logical: If TRUE (default is FALSE), the program will remove high temperature data points that significantly deviate from a straight line fit. This can reduce false positives but also true positives. Use with caution.}

\item{normalization.points}{Integer: How many points should be used for baseline normalization of high temperature data points? Default is 1, but 2 are recommended for 10 doses and even more for more doses, depending on the thermal behaviour of the studied system.}

\item{normalize.baseline}{Logical: If TRUE (default), all points of baseline data (lowest temperature) will be used to normalize the baseline and the average position of all points will be 1. If FALSE, the number of points given in 'normalization.points' will be used.}

\item{normalize.selection}{Vector of length 2: If the data should be normalized only on a subset of data, what column and value should be used? For example, 
'normalize.selection=c("Organism","Pf")' will normalized only based on the data that have value 'Pf' in the column 'Organism'. Default is NULL.}
}
\value{
A list with three elements. $data is a data frame with  the scaled data, $all_data is a data.frame with the input and output data and $plot is a ggplot with scaling boxplots.
}
\description{
Scaling and normalization of the data.
}
\examples{
data_scaled <- ITDRMS.scale(data_cleaned)
}
