% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pf_meltcurve_incell.R, R/Pf_meltcurve_lysate.R
\docType{data}
\name{Pf_meltcurve_lysate}
\alias{Pf_meltcurve_lysate}
\title{Plasmodium falciparum intact cell melting curves}
\format{
An object of class \code{data.frame} with 5283 rows and 11 columns.

An object of class \code{data.frame} with 5283 rows and 11 columns.
}
\usage{
data(Pf_meltcurve_incell)

data(Pf_meltcurve_lysate)
}
\description{
The melting curves were recorded at 34 hpi synchronized culture
in intact cell setup./

The melting curves were recorded at 34 hpi synchronized culture
in lysate setup.
}
\examples{
data(Pf_meltcurve_incell)
ITDRMS.POS("O14818",Pf_meltcurve_incell)
data(Pf_meltcurve_lysate)
ITDRMS.POS("O14818",Pf_meltcurve_lysate)
}
\references{
Pazicky et al. (2024) Unpublished
(\href{tbd}{PubMed})

Pazicky et al. (2024) Unpublished
(\href{tbd}{PubMed})
}
\keyword{datasets}
