% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITDRMS.hit2.R
\name{ITDRMS.hit2}
\alias{ITDRMS.hit2}
\title{ITDRMS.hit2}
\usage{
ITDRMS.hit2(
  data = NULL,
  minresponse = c(1, 0.5),
  R2line = 0.6,
  POI = c(),
  plot.settings = list()
)
}
\arguments{
\item{data}{Data frame: Scaled data with removed outliers and fitting statistics, ideally $data element from ITDRMS.fit output.}

\item{minresponse}{Integer vector: Minimal change in the solubility to be considered hit or candidate. Default is c(1,0.5).}

\item{R2line}{Double: R2 cut-off. Proteins with R-squared value below this value will not be considered as hits or candidates despite favorable response and confidence index.}

\item{POI}{Character vector: ID of the protein or proteins to be highlighted on the plot.
\preformatted{
list(
  labels = TRUE,
  label.text.size = 2.5,
  label.force = 1.3,
  xlims = c(-max(1, abs(hit_data$total.response)) - 1, 
            max(1, abs(hit_data$total.response)) + 1),
  ylims = c(min(-log10(hit_data$CI)), max(-log10(hit_data$CI))),
  point.sizes = c(1, 2),
  point.colors = c("gray", "red3", "green3", "coral2", "darkolivegreen2"),
  axis.title.size = 18,
  axis.text.size = 16,
  legend.position = "bottom",
  legend.text.size = 8,
  legend.title = element_blank(),
  POI.color = "blue",
  POI.size = 3.5
)
}}
}
\value{
A list with three elements. $data is a data frame with dAUC and p-values for all proteins, $plot is the volcano plot based on the hits and 
$hitlist is a list with two vectors: one for Stabilized and one for Destabilized hits.
}
\description{
Identifies hits from fitted mass spec data.
}
\examples{
data_fitted <- ITDRMS.hit(data_fitted, nMAD=0.2)
}
